### 2021-11-19 #152 @whedon
[PRE REVIEW]: An open learning resource on Reproducible Data Science with Open-Source Python Tools and Real-World Data
**Submitting author:** @valdanchev (<a href="http://orcid.org/0000-0002-7563-0168">Valentin Danchev</a>)
**Repository:** <a href="https://github.com/valdanchev/reproducible-data-science-python" target ="_blank">https://github.com/valdanchev/reproducible-data-science-python</a>
**Version:** v1.0.0
**Managing EiC:** Jordan Gorzalski
