+++ paper.bib
@@ -100,23 +100,20 @@
 
 @article{Hagberg2008networkx,
   title = {Exploring network structure, dynamics, and function using networkx},
-  author = {Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
-  doi = {},
-  url = {https://www.osti.gov/biblio/960616}, 
-  journal = {},
-  number = ,
-  volume = ,
-  place = {United States},
+  author = {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
+  url = {http://conference.scipy.org/proceedings/scipy2008}, 
+  booktitle =   {Proceedings of the 7th Python in Science Conference (SciPy 2008)},
+  pages = {11-16},
+  place = {Pasadena, CA, United States},
   year = {2008},
-  month = {1},
+  editor =    {Gael Varoquaux and Travis Vaught and Jarrod Millman},
 }
 
 @article{Pedregosa2011scikit, 
   title={Scikit-learn: Machine learning in Python}, 
-  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others}, 
+  author={Pedregosa, Fabian and Varoquaux, Gael and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others}, 
   journal={Journal of machine learning research}, 
-  volume={12}, 
-  number={Oct}, 
+  volume={12},
   pages={2825--2830}, 
   year={2011},
 }
