+++ paper.bib
@@ -101,9 +101,9 @@
 @article{Hagberg2008networkx,
   title = {Exploring network structure, dynamics, and function using networkx},
   author = {Aric A. Hagberg and Daniel A. Schult and Pieter J. Swart},
-  url = {http://conference.scipy.org/proceedings/scipy2008}, 
+  url = {http://conference.scipy.org/proceedings/SciPy2008/paper_2}, 
   booktitle =   {Proceedings of the 7th Python in Science Conference (SciPy 2008)},
-  pages = {11-16},
+  pages = {11-15},
   place = {Pasadena, CA, United States},
   year = {2008},
   editor =    {Gael Varoquaux and Travis Vaught and Jarrod Millman},
