+++ paper.md
@@ -11,16 +11,14 @@ tags:
 - social networks
 - data science ethics
 date: "19 November 2021"
-output: pdf_document
-urlcolor: blue
 authors:
 - name: Valentin Danchev
   orcid: 0000-0002-7563-0168
   affiliation: 1
-bibliography: paper.bib
 affiliations:
 - name: Department of Sociology, University of Essex, United Kingdom
   index: 1
+bibliography: paper.bib
 ---
 
 # Summary
