### 2021-11-24 valdanchev/reproducible-data-science-python#2 @mateusbraga
Issue on page /prediction_supervised_learning.html
There is an `ModuleNotFoundError: No module named 'google'` error being shown in the 'Accessing data from your Google Drive' session.
