### 2021-11-29 #152 @valdanchev
Hi @jgorzalski, I have corrected the set of problems I've got during submission and the PDF seems to compile now. Looking forward to hearing from you about next steps.
