### 2021-11-29 #152 @jgorzalski
Hello @valdanchev 👋 Thanks for bringing this to my attention! I had missed your correction of the error and am just getting back to it from the short Thanksgiving break here in the USA. I will take a look at this and try to get an editor assigned promptly. After that the editor will begin the search for reviewers. Have you had a chance to look at our list of reviewers (linked in the first comment from whedon as "list of people") or do you have any suggestions of people who could be potential reviewers?
