### 2021-12-13 #152 @ShanEllis
Hi @mnarayan, reaching out to see if you would be interested in reviewing this submission titled "_An open learning resource on Reproducible Data Science with Open-Source Python Tools and Real-World Data_" for JOSE, as I think it would be a good fit for your interests and expertise.

I know it's a busy time of year for many, so please do feel free to let me know if you're interested but may need a bit longer (into the New year) to review.

Finally, if you have any questions about the open-review process with JOSE, please feel free to ask away!

Thanks!
Shannon
