### 2021-12-14 #152 @ShanEllis
Just adding a note that Tom and I chatted about the fact that he will not be able to get to this review until January (which I think makes total sense!). Just wanted to include that in the discussion here, so everyone was on the same page and aware of expected timeline.
