### 2022-01-12 #152 @ShanEllis
> So sorry, but I am not available for reviews during January.
> […](#)
> On Mon, Dec 13, 2021 at 3:00 PM Shannon Ellis ***@***.***> wrote: Hi @mnarayan <https://github.com/mnarayan>, reaching out to see if you would be interested in reviewing this submission titled "*An open learning resource on Reproducible Data Science with Open-Source Python Tools and Real-World Data*" for JOSE, as I think it would be a good fit for your interests and expertise. I know it's a busy time of year for many, so please do feel free to let me know if you're interested but may need a bit longer (into the New year) to review. Finally, if you have any questions about the open-review process with JOSE, please feel free to ask away! Thanks! Shannon — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub <[#152 (comment)](https://github.com/openjournals/jose-reviews/issues/152#issuecomment-992988297)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AAFC753XXGJR6GS7RWRDDXDUQZ3JTANCNFSM5IM4BEWA> . Triage notifications on the go with GitHub Mobile for iOS <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675> or Android <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.

@mnarayan  I totally understand! Thanks for the reply!
