### 2022-01-12 #152 @ShanEllis
Hi @lechten - I'm reaching out to see if you're interested in reviewing a manuscript for JOSE titled "An open learning resource on Reproducible Data Science with Open-Source Python Tools and Real-World Data", as I think it would be a good fit for your interests/expertise and passion for OER.

And, if you have any questions about the open-review process with JOSE/making your decision, please feel free to ask away!
