### 2022-01-12 #156 @whedon
[REVIEW]: An open learning resource on Reproducible Data Science with Open-Source Python Tools and Real-World Data
**Submitting author:** @valdanchev (<a href="http://orcid.org/0000-0002-7563-0168">Valentin Danchev</a>)
**Repository:** <a href="https://github.com/valdanchev/reproducible-data-science-python" target ="_blank">https://github.com/valdanchev/reproducible-data-science-python</a>
**Version:** v2.1.2

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
