### 2022-01-25 valdanchev/reproducible-data-science-python#6 @TomDonoghue
JOSE Review: repository organization & ordering notebooks
This issue / suggestion is related to the [JOSE review](https://github.com/openjournals/jose-reviews/issues/156).

I would recommend potentially organizing the files in the repository a bit more, with two particular goals:
- currently, the order and organization of materials isn't particularly obvious
  - for example, if one just follows the "Binder" or "Collab" links from the README link, it's not clear which notebooks to start with
  - In particular, following the Binder link, we see an un-ordered list of all files, and I don't think it's clear how to proceed / interact with these files in a sensible ordering, if one wants to work through the materials interactively
- for anyone looking to use / deploy this resource in a course, the organization of materials isn't very obvious
  - I think it would be easier for a potential teacher to see what's what by having organized sub-folders of different types of things

Related notes:
- there are Markdown files that I think come from the JupyterBook tempate (ex `content.md`, `markdown.md`), and I think are not used. Can these be deleted?
- content files could be organized into their own sub-folder
  - for example, for images, in this kind of repo, I tend to use an `img/` sub-folder
  - overall, you could use something like `img/`, `data/`, `markdown/` sub-folders for different aspects of the content
- I think the JOSE paper could also be moved to a sub-folder, so it's clear this is separate from the textbook content
  - this would also help to group related files together
- assuming there is an explicit numbering to the notebooks, using leading numbers (eg `01`, `02`) in the titles would automatically sort these files, and make the order clear
- the `.DS_Store` and `.ipynb_checkpoints` files can presumably be deleted

Some of this file movements might require checking file paths for building and deploying the book, but I think it should all work.
