### 2022-01-25 #156 @TomDonoghue
I've worked through most of my review, including exploring through the content, and checking I can run some examples through Binder etc.

I have some practical suggestions, that I have described in detail on the resource repository, including:
- [a suggestion to organize and more explicitly order the content](https://github.com/valdanchev/reproducible-data-science-python/issues/6)
- [some minor formatting and content notes for the notebooks](https://github.com/valdanchev/reproducible-data-science-python/issues/5)

Other notes:
- In terms of requirements, I think there is an implicit dependency whereby some notebooks only work on Collab (not Binder), because they load the `google` module. This is captured in this [pre-existing issue](https://github.com/valdanchev/reproducible-data-science-python/issues/2). I don't think this is a huge issue, per se, but I think some notes in the notebooks about this kind of extra dependency (what works where and/or what would need to be installed) could be added somewhere.
- There is not a version identifier that I found, though I admit I'm not entirely sure how versions are supposed to work for this kind of resource
- Though it's perhaps implicit, for Community Guidelines, an explicit note in the README, and/or textbook landing page about what to do to make a suggestion / report a problem (open an issue, presumably) might be useful, especially for students who might not be clear on standard Github practices.

In terms of the literal content - I don't have as much to say. Based on skimming it all, I think all the content (ideas and concepts) look good to me. It's a bit tricky for me to know exactly how an unfamiliar student would feel about the progression through things. My sense is perhaps that there is a lot introduced quite quickly, and one might need a bit of a sense of basic coding to really be able to jump in - but under the idea of it as a "crash course", I think it broadly works, and is a useful resource to have available.

My one more practical content note is that I think the `end_to_end_data_science_project`, which in particular feels like jumping in the deep end, could be more explicit that it skims through topics as an introduce (it effectively overviews the course), and that these topics are dealt with in more detail in future notebooks, and that it might make sense for students "jump ahead" to practice these topics there first (at first I even thought maybe this notebook was supposed to come at the end?).

Overall, I think this looks like an interesting resource, with it's strengths being that it:
- introduces a lot of topics in data science, with a focus on practical elements (reproducibility, ethics, e
- has a lot of links to other resources to map out other places to learn things
- uses some real data to demonstrate these things in practice, and is re-runnable by students
- is focused at a particular area / group of students (social science / sociology) for whom there be fewer dedicated materials

Conclusion: I think if the relatively minor notes I've mentioned can be fixed up a bit, my overall impression is that this is a useful resource, responsive to JOSE's aims and requirements, and after the review edits I'd be happy to sign off on it as a reviewer.
