### 2022-01-27 valdanchev/reproducible-data-science-python#5 @valdanchev
Many thanks for the very helpful feedback @TomDonoghue. Will go through your comments and will write back when I manage to address them.
