### 2022-01-27 valdanchev/reproducible-data-science-python#6 @valdanchev
Agree @TomDonoghue, the resource repository could be better organised indeed. Your thoughtful suggestions will help me greatly to improve the structure of the content.
