### 2022-01-27 #156 @valdanchev
Thank you so much for these very helpful comments @TomDonoghue. Will go through all of your notes and issues and will write back when I have a revised version that addresses them.
