### 2022-02-14 valdanchev/reproducible-data-science-python#7 @lechten
JOSE Review: Typos, URLs
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/156).

- In Acknowledgements, a closing parenthesis is missing (after “(BITSS)”?).
- L. 96, “The resource is designed to scaffold reproducible and transparent research workflow”: Should this be “a” workflow or “workflows”?
- The checklist asks for DOIs.  Not all papers have one.  Did you check?  Note that the URL in Lau et al. is not formatted as hyperlink.
- The URL in Guzharina is too long.  It should be broken, maybe at “-”.
