### 2022-02-14 valdanchev/reproducible-data-science-python#8 @lechten
JOSE Review: Code quality
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/156).

One of the learning objectives is: “Write clean, reusable, and reproducible code in Python.”
For “clean code”, I strongly suggest to add pointers for style guides and tools.  Then, please run pylint or flake8 on your notebooks, e.g., with [nbQA](https://github.com/nbQA-dev/nbQA), and decide what rules to follow.
In what regard is “reusable” addressed?
