### 2022-02-14 valdanchev/reproducible-data-science-python#9 @lechten
JOSE Review: Choice of and pointers to Jupyter environments
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/156).

In Section “Getting started”, MyBinder and Colab are mentioned but their differences remain vague.  Why should students choose what variant?  Afterwards, only links to Colab are provided.  Why?  For the purposes of this paper, a single link to the online textbook might be sufficient.  The textbook could then offer clear guidance concerning the question what notebooks to open when (which is related to issue #6).

Note that on https://valdanchev.github.io/reproducible-data-science-python/end_to_end_data_science_project.html the link [Google Colaboratory](https://valdanchev.github.io/reproducible-data-science-python/(https://colab.research.google.com/notebooks/intro.ipynb#recent=true)) is broken.  I’m confused where students are supposed to switch from the textbook to Jupyter environments.  I suggest to add explicit pointers in that regard.

In addition, “Python for Data Analysis on the Cloud” also mentions this: “Students with no Google account will have an opportunity to access the Jupyter notebook on the cloud via JupyterHub and The Littlest JupyterHub (also known as TLJH) in particular.”
I do not know what that means.

As the resources also aims to teach ethics issues, I suggest to clarify Colab terms of service: What is tracked by Google for what purposes?
