### 2022-02-14 valdanchev/reproducible-data-science-python#10 @lechten
JOSE Review: Learning resources for data ethics?
This is related to your [JOSE submission](https://github.com/openjournals/jose-reviews/issues/156).

The notebook on data ethics mentions three resources under “Learning resources about the A-level controversy”.  Please add hyperlinks.
