### 2022-02-25 #156 @valdanchev
Thanks so much, @lechten for your very helpful comments and suggestions for improvements. The pointer to nbqa is extremely useful. My teaching-heavy term will be over in a few weeks, and I am looking forward to then revising the resource based on your and @TomDonoghue comments.
