+++ paper.bib
@@ -1,205 +1,0 @@
-@article{Adhikari2021,
-  journal = {Harvard Data Science Review},
-  doi = {10.1162/99608f92.cb0fa8d2},
-  note = {https://hdsr.mitpress.mit.edu/pub/e69066t4},
-  publisher = {},
-  title = {Interleaving Computational and Inferential Thinking: Data Science for Undergraduates at Berkeley},
-  url = {https://hdsr.mitpress.mit.edu/pub/e69066t4},
-  author = {Adhikari, Ani and DeNero, John and Jordan, Michael I.},
-  date = {2021-04-30},
-  year = {2021},
-  month = {4},
-  day = {30},
-}
-
-@BOOK{NAP2018,
-  author    = {National Academies of Sciences, Engineering, and Medicine},
-  title     = {Data Science for Undergraduates: Opportunities and Options},
-  isbn      = {978-0-309-47559-4},
-  doi       = {10.17226/25104},
-  url       = {https://www.nap.edu/catalog/25104/data-science-for-undergraduates-opportunities-and-options},
-  year      = {2018},
-  publisher = {The National Academies Press},
-  address   = {Washington, DC},
-}
-
-@inproceedings{Danyluk2021, 
-  author = {Danyluk, Andrea and Leidig, Paul and McGettrick, Andrew and Cassel, Lillian and Doyle, Maureen and Servin, Christian and Schmitt, Karl and Stefik, Andreas}, 
-  title = {Computing Competencies for Undergraduate Data Science Programs: An ACM Task Force Final Report}, year = {2021}, isbn = {9781450380621}, 
-  publisher = {Association for Computing Machinery}, 
-  address = {New York, NY, USA}, url = {https://doi.org/10.1145/3408877.3432586}, doi = {10.1145/3408877.3432586}, 
-  booktitle = {Proceedings of the 52nd ACM Technical Symposium on Computer Science Education}, 
-  pages = {1119–1120},
-  numpages = {2},
-  keywords = {data science, curriculum},
-  location = {Virtual Event, USA}, 
-  series = {SIGCSE '21},
-}
-
-@article{Arnold2019turing,
-  title={The turing Way: a handbook for reproducible data science},
-  author={Arnold, Becky and Bowler, Louise and Gibson, Sarah and Herterich, Patricia and Higman, Rosie and Krystalli, Anna and Morley, Alexander and O'Reilly, Martin and Whitaker, Kirstie and others},
-  journal={Zenodo},
-  year={2019},
-}
-
-@article{Stoudt2021,
-  doi = {10.1371/journal.pcbi.1008770},
-  author = {Stoudt, Sara AND Vásquez, Váleri N. AND Martinez, Ciera C.},
-  journal = {PLOS Computational Biology},
-  publisher = {Public Library of Science},
-  title = {Principles for data analysis workflows},
-  year = {2021},
-  month = {03},
-  volume = {17},
-  url = {https://doi.org/10.1371/journal.pcbi.1008770},
-  pages = {1-26},
-  number = {3},
-}
-
-@conference{Kluyver2016jupyter, 
-  Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows}, 
-  Author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing}, 
[…]
