+++ notebooks/requirements.txt
@@ -1,0 +1,1 @@
+pandas==1.1.3
