+++ _sources/README.md
@@ -1,4 +1,60 @@
-# reproducible-data-science-python
-Reproducible Data Science: Accessible Data Analysis with Open Source Python Tools and Real-World Data
+# Reproducible Data Science with Python
 
-[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/master)
+This repository contains the open learning resource [Reproducible Data Science with Python](https://valdanchev.github.io/reproducible-data-science-python/intro.html) in the form of Python Jupyter notebooks.
+
+## Description 
+
+The open learning resource uses real-world social data sets related to the COVID-19 pandemic to provide an accessible introduction to open, reproducible, and ethical data analysis using hands-on Python coding, modern open-source computational tools, and data science techniques. 
+Topics include reproducible workflows, data wrangling, exploratory data analysis, data visualisation, pattern discovery (e.g., clustering), prediction and machine learning, causal inference, and network analysis.
+ 
+## How to use the learning resource?
+
+You can read the textbook on the dedicated [website](https://valdanchev.github.io/reproducible-data-science-python/intro.html). In addition, you can view each individual notebook on GitHub by clicking on the respective ![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label) button below.
+
+To interactively work with the code, you can access the interactive versions of the Jupyter notebooks via the free cloud services [myBinder](https://mybinder.org) and [Colab](https://colab.research.google.com/notebooks/intro.ipynb#recent=true). Both services allow you to interactively modify and run the notebooks from your browser. 
+
+By clicking on a ![Binder](https://mybinder.org/badge_logo.svg) button below, you will launch an interactive version of the Jupyter notebook, with the following capabilities:
+* **Reproducibility:** Notebooks run in a reproducible computing environment containing the same Python packages and package versions used in the original notebooks.
+* **Session time:** Notebooks run for up to 6 hours and will be shut down automatically after more than 10 minutes of inactivity.
+* **Notebook persistence:** Non-persistent, changes will be lost after your myBinder session times out unless you download the notebook.
+* **Access:** Free, public, and anonymous cloud service. No setup or a login is required to view and execute the notebooks.
+
+By clicking on a ![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg) button below, you will open a Jupyter notebook in Colab, with the following capabilities:
+* **Reproducibility:** Colab environment comes with pre-installed Python packages and package versions, which may differ from the ones used in the original notebooks. To enable computational reproducibility, see section "Installing dependencies" below.
+* **Session time:** Notebooks run for up to 12 hours and will disconnect when left idle for too long (time may vary).
+* **Notebook persistence:** Persistent, changes are saved automatically when you are logged in.
+* **Access:** Free cloud service that requires no setup. You can view the notebooks without a login but to execute and modify a notebook, a Google account and a login are required.
+
+| Textbook chapter | View on GitHub | Launch on myBinder.org | Open in Colab |
+|--------------|-----------|--------------|-----------|
[…]
