+++ _sources/intro.md
@@ -1,13 +1,12 @@
-Welcome to Reproducible Data Science
+Reproducible Data Science with Python
 =======================
 
-Welcome to the textbook ***Reproducible Data Science with Open-Source Python Tools and 
-Real-World Data*** by Valentin Danchev. The textbook uses real-world 
+Welcome to the textbook ***Reproducible Data Science with Python*** by Valentin Danchev. The textbook uses real-world 
 social data sets related to the COVID-19 pandemic to provide an accessible introduction to 
 open, reproducible, and ethical data analysis using hands-on Python coding, modern open-source 
 computational tools, and data science techniques. Topics include open reproducible research workflows, 
 data wrangling, exploratory data analysis, data visualisation, pattern discovery (e.g., clustering), 
-prediction and machine learning, causal inference, and network analysis.
+prediction & machine learning, causal inference, and network analysis.
 
 ## Contact
 
@@ -23,4 +22,4 @@ University of Essex\
 
 [![License: CC BY-SA 4.0](https://img.shields.io/badge/License-CC%20BY--SA%204.0-lightgrey.svg)](http://creativecommons.org/licenses/by-sa/4.0/)
 
-_Reproducible Data Science with Open-Source Python Tools and Real-World Data_ by [Valentin Danchev](https://valdanchev.github.io) is licensed under a [Creative Commons Attribution-ShareAlike 4.0 International License](https://creativecommons.org/licenses/by-sa/4.0/).
+_Reproducible Data Science with Python_ by [Valentin Danchev](https://valdanchev.github.io) is licensed under a [Creative Commons Attribution-ShareAlike 4.0 International License](https://creativecommons.org/licenses/by-sa/4.0/).
