+++ README.md
@@ -28,16 +28,16 @@ By clicking on a ![Open In Colab](https://colab.research.google.com/assets/colab
 | Textbook chapter | View on GitHub | Launch on myBinder.org | Open in Colab |
 |--------------|-----------|--------------|-----------|
 | [About the textbook](https://valdanchev.github.io/reproducible-data-science-python/notebooks/00_textbook_outline.html) | [![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label)](https://github.com/valdanchev/reproducible-data-science-python/blob/master/notebooks/00_textbook_outline.ipynb) | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/master) | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/valdanchev/reproducible-data-science-python/blob/master/notebooks/00_textbook_outline.ipynb)  |
-| [End-to-End Data Science Project](https://valdanchev.github.io/reproducible-data-science-python/notebooks/01_end_to_end_data_science_project.html) | [![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label)]() | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/master) | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/valdanchev/reproducible-data-science-python/blob/master/notebooks/01_end_to_end_data_science_project.ipynb) |
-| [Python Data Science on the Cloud](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html) | [![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label)]() | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/master) | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/valdanchev/reproducible-data-science-python/blob/master/notebooks/02_python_data_science_on_the_cloud.ipynb) |
-| [Open Reproducible Data Science Workflow](https://valdanchev.github.io/reproducible-data-science-python/notebooks/03_open_reproducible_workflows.html) | [![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label)]() | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/master) | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/valdanchev/reproducible-data-science-python/blob/master/notebooks/03_open_reproducible_workflows.ipynb) |
-| [Data Design and Data Wrangling](https://valdanchev.github.io/reproducible-data-science-python/notebooks/04_data_design_and_data_wrangling.html) | [![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label)]() | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/master) | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/valdanchev/reproducible-data-science-python/blob/master/notebooks/04_data_design_and_data_wrangling.ipynb) |
[…]
