+++ README.md
@@ -7,6 +7,11 @@ This repository contains the open learning resource [Reproducible Data Science w
 The open learning resource uses real-world social data sets related to the COVID-19 pandemic to provide an accessible introduction to open, reproducible, and ethical data analysis using hands-on Python coding, modern open-source computational tools, and data science techniques. 
 Topics include reproducible workflows, data wrangling, exploratory data analysis, data visualisation, pattern discovery (e.g., clustering), prediction and machine learning, causal inference, and network analysis.
  
+| **Publications About the Learning Resource** | [![status](https://jose.theoj.org/papers/3e1de7c74161a5b2c4ce74e536ef6898/status.svg)](https://jose.theoj.org/papers/3e1de7c74161a5b2c4ce74e536ef6898)[![DOI](https://jose.theoj.org/papers/10.21105/jose.00152/status.svg)](https://doi.org/10.21105/jose.00152) |
+| :--- | :--- |
+| **Releases** | [![GitHub release (latest by date)](https://img.shields.io/github/v/release/valdanchev/reproducible-data-science-python?logo=GitHub)](https://github.com/valdanchev/reproducible-data-science-python/releases) |
+| __License__ | [![License: CC BY-SA 4.0](https://img.shields.io/badge/License-CC%20BY--SA%204.0-lightgrey.svg)](http://creativecommons.org/licenses/by-sa/4.0/) |
+
 ## How to use the learning resource?
 
 You can read the textbook on the dedicated [website](https://valdanchev.github.io/reproducible-data-science-python/intro.html). In addition, you can view each individual notebook on GitHub by clicking on the respective ![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label) button below.
