+++ README.md
@@ -16,13 +16,13 @@ Topics include reproducible workflows, data wrangling, exploratory data analysis
 
 You can read the textbook on the dedicated [website](https://valdanchev.github.io/reproducible-data-science-python/intro.html). In addition, you can view each individual notebook on GitHub by clicking on the respective ![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label) button below.
 
-To interactively work with the code, you can access the interactive versions of the Jupyter notebooks via the free cloud services [myBinder](https://mybinder.org) and [Colab](https://colab.research.google.com/notebooks/intro.ipynb#recent=true). Both services allow you to interactively modify and run the notebooks from your browser. 
+To interactively work with the code, you can access the interactive versions of the Jupyter notebooks via the free cloud services [MyBinder](https://mybinder.org) and [Colab](https://colab.research.google.com/notebooks/intro.ipynb#recent=true). Both services allow you to interactively modify and run the notebooks from your browser. 
 
 By clicking on a ![Binder](https://mybinder.org/badge_logo.svg) button below, you will launch an interactive version of the Jupyter notebook, with the following capabilities:
 * **Reproducibility:** Notebooks run in a reproducible computing environment containing the same Python packages and package versions used in the original notebooks.
 * **Session time:** Notebooks run for up to 6 hours and will be shut down automatically after more than 10 minutes of inactivity.
-* **Notebook persistence:** Non-persistent, changes will be lost after your myBinder session times out unless you download the notebook.
-* **Access:** Free, public, and anonymous cloud service. No setup or a login is required to view and execute the notebooks.
+* **Notebook persistence:** Non-persistent, changes will be lost after your MyBinder session times out unless you download the notebook.
+* **Access:** Free, public, and anonymous cloud service. No setup or a login is required to view and execute the notebooks. Notebooks that use safeguarded data should not be launched on MyBinder.
 
 By clicking on a ![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg) button below, you will open a Jupyter notebook in Colab, with the following capabilities:
 * **Reproducibility:** Colab environment comes with pre-installed Python packages and package versions, which may differ from the ones used in the original notebooks. To enable computational reproducibility, see section "Installing dependencies" below.
@@ -30,7 +30,7 @@ By clicking on a ![Open In Colab](https://colab.research.google.com/assets/colab
 * **Notebook persistence:** Persistent, changes are saved automatically when you are logged in.
 * **Access:** Free cloud service that requires no setup. You can view the notebooks without a login but to execute and modify a notebook, a Google account and a login are required.
 
-| Textbook chapter | View on GitHub | Launch on myBinder.org | Open in Colab |
+| Textbook chapter | View on GitHub | Launch on MyBinder.org | Open in Colab |
 |--------------|-----------|--------------|-----------|
[…]
