+++ README.md
@@ -43,8 +43,9 @@ By clicking on a ![Open In Colab](https://colab.research.google.com/assets/colab
 | [What Causes What? Introduction to Causal inference](https://valdanchev.github.io/reproducible-data-science-python/notebooks/08_causal_inference.html) | [![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label)](https://github.com/valdanchev/reproducible-data-science-python/blob/master/notebooks/08_causal_inference.ipynb) | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/604f6843faf5052420d4254073e0ea1db809864c?urlpath=lab%2Ftree%2Fnotebooks%2F04_data_design_and_data_wrangling.ipynb) | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/valdanchev/reproducible-data-science-python/blob/master/_build/html/_sources/notebooks/08_causal_inference.ipynb) |
 | [Network Analysis](https://valdanchev.github.io/reproducible-data-science-python/notebooks/09_network_analysis.html) | [![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label)](https://github.com/valdanchev/reproducible-data-science-python/blob/master/notebooks/09_network_analysis.ipynb) | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/604f6843faf5052420d4254073e0ea1db809864c?urlpath=lab%2Ftree%2Fnotebooks%2F04_data_design_and_data_wrangling.ipynb) | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/valdanchev/reproducible-data-science-python/blob/master/_build/html/_sources/notebooks/09_network_analysis.ipynb) |
 | [Data Ethics](https://valdanchev.github.io/reproducible-data-science-python/notebooks/10_data_ethics.html) | [![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label)](https://github.com/valdanchev/reproducible-data-science-python/blob/master/notebooks/10_data_ethics.ipynb) | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/valdanchev/reproducible-data-science-python/604f6843faf5052420d4254073e0ea1db809864c?urlpath=lab%2Ftree%2Fnotebooks%2F04_data_design_and_data_wrangling.ipynb) | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/valdanchev/reproducible-data-science-python/blob/master/_build/html/_sources/notebooks/10_data_ethics.ipynb) |
-_Note:_The notebooks _Prediction using Supervised Learning_ and _What Causes What? Introduction to Causal inference_ require [access to safeguarded data](https://valdanchev.github.io/reproducible-data-science-python/notebooks/07_prediction_using_supervised_learning.html#data-understanding-society-covid-19) which, once obtained, needs to be stored securely on your Google Drive and the notebooks should be run on your private Colab notebooks.  
 <br>
+_Note:_ The notebooks _Prediction using Supervised Learning_ and _What Causes What? Introduction to Causal inference_ require [access to safeguarded data](https://valdanchev.github.io/reproducible-data-science-python/notebooks/07_prediction_using_supervised_learning.html#data-understanding-society-covid-19) which, once obtained, needs to be stored securely on your Google Drive and the notebooks should be run on your private Colab notebooks.  
+
 
 ## Installing dependencies
 
