### 2022-05-18 valdanchev/reproducible-data-science-python#5 @valdanchev
In relation to the notes on formatting, the code now conforms to PEP 8 by running automated code formatters *Black* and *flake8* on Jupyter through [nbQA](https://github.com/nbQA-dev/nbQA).
* Added a section [Readability of your Python code](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#readability-of-your-python-code)

The printing of long elements was reduced by:
* enabling scrolling for outputs such that outputs are not listed in long segments but are put into scrollable frames;
* printing the full output in a .txt file (e.g., requirements.txt).

To address the notes on content, added the following sections in the notebook 'Python for Data Analysis on the Cloud':
* [Running Terminal commands in Jupyter/Colab notebook](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#running-terminal-commands-in-jupyter-colab-notebook)
* [Installing packages with pip](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#installing-packages-with-pip)
* [Jupyter magic commands](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#jupyter-magic-commands)
The three sections describe and give examples of relevant command-line tools and syntax different from Python.
