### 2022-05-18 valdanchev/reproducible-data-science-python#6 @valdanchev
Following the very helpful suggestions, the files are now reorganised in the repository and ordered in a structured way:
* different types of content are now assigned to sub-folders `notebooks/`, `images/`, `data/`, `JOSE_paper/`;
* notebooks are sorted in ascending order in folder `notebooks/`
* unused/unnecessary files are deleted

As a result, the organisation of the [repository for the learning resource](https://github.com/valdanchev/reproducible-data-science-python), including notebooks' sequential order, looks much more intuitive and structured.
