### 2022-05-18 valdanchev/reproducible-data-science-python#7 @valdanchev
Thank you for the helpful comments, which I have addressed:
* corrected the typos
* added DOIs for almost all papers (except for Hagberg et al which does not seem to be available)
* formated the URL in Guzharina to fit
