### 2022-05-18 valdanchev/reproducible-data-science-python#8 @valdanchev
Added a section entitled [Readability of your Python code](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#readability-of-your-python-code) in which PEP 8 style guide for Python code is introduced and key relevant recommendations for writing a readable Python code are outlined, including (i) naming conventions, (ii) maximum line length, (iii) indentation, (iv) tabs or spaces?, (v) whitespace in expressions and statements.

The resource now conforms to the Python’s official [PEP 8 style guide for Python code](https://peps.python.org/pep-0008/)—the automated code formatter [Black](https://github.com/psf/black) and style guide checker [flake8](https://flake8.pycqa.org/en/latest/) were executed via [nbQA](https://github.com/nbQA-dev/nbQA) on all Jupyter notebooks. The notebook performing the formatting is included in the repository folder [`tools`](https://github.com/valdanchev/reproducible-data-science-python/tree/master/tools).

A new section on code reusability named [Code reuse: creating your own functions](https://valdanchev.github.io/reproducible-data-science-python/notebooks/05_data_exploration_and_visualisation.html#code-reuse-creating-your-own-functions) was added.
