+++ _sources/README.md
@@ -2,6 +2,11 @@
 
 This repository contains the open learning resource [Reproducible Data Science with Python](https://valdanchev.github.io/reproducible-data-science-python/intro.html) in the form of Python Jupyter notebooks.
 
+| **Publication about the Learning Resource** | [![status](https://jose.theoj.org/papers/3e1de7c74161a5b2c4ce74e536ef6898/status.svg)](https://jose.theoj.org/papers/3e1de7c74161a5b2c4ce74e536ef6898)[![DOI](https://jose.theoj.org/papers/10.21105/jose.00152/status.svg)](https://doi.org/10.21105/jose.00152) |
+| :--- | :--- |
+| **Releases** | [![GitHub release (latest by date)](https://img.shields.io/github/v/release/valdanchev/reproducible-data-science-python?logo=GitHub)](https://github.com/valdanchev/reproducible-data-science-python/releases) |
+| __License__ | [![License: CC BY-SA 4.0](https://img.shields.io/badge/License-CC%20BY--SA%204.0-lightgrey.svg)](http://creativecommons.org/licenses/by-sa/4.0/) |
+
 ## Description 
 
 The open learning resource uses real-world social data sets related to the COVID-19 pandemic to provide an accessible introduction to open, reproducible, and ethical data analysis using hands-on Python coding, modern open-source computational tools, and data science techniques. 
@@ -11,13 +16,13 @@ Topics include reproducible workflows, data wrangling, exploratory data analysis
 
 You can read the textbook on the dedicated [website](https://valdanchev.github.io/reproducible-data-science-python/intro.html). In addition, you can view each individual notebook on GitHub by clicking on the respective ![GitHub](https://badgen.net/badge/icon/GitHub?icon=github&label) button below.
 
-To interactively work with the code, you can access the interactive versions of the Jupyter notebooks via the free cloud services [myBinder](https://mybinder.org) and [Colab](https://colab.research.google.com/notebooks/intro.ipynb#recent=true). Both services allow you to interactively modify and run the notebooks from your browser. 
+To interactively work with the code, you can access the interactive versions of the Jupyter notebooks via the free cloud services [MyBinder](https://mybinder.org) and [Colab](https://colab.research.google.com/notebooks/intro.ipynb#recent=true). Both services allow you to interactively modify and run the notebooks from your browser. 
 
 By clicking on a ![Binder](https://mybinder.org/badge_logo.svg) button below, you will launch an interactive version of the Jupyter notebook, with the following capabilities:
 * **Reproducibility:** Notebooks run in a reproducible computing environment containing the same Python packages and package versions used in the original notebooks.
 * **Session time:** Notebooks run for up to 6 hours and will be shut down automatically after more than 10 minutes of inactivity.
-* **Notebook persistence:** Non-persistent, changes will be lost after your myBinder session times out unless you download the notebook.
-* **Access:** Free, public, and anonymous cloud service. No setup or a login is required to view and execute the notebooks.
+* **Notebook persistence:** Non-persistent, changes will be lost after your MyBinder session times out unless you download the notebook.
+* **Access:** Free, public, and anonymous cloud service. No setup or a login is required to view and execute the notebooks. Notebooks that use safeguarded data should not be launched on MyBinder.
 
[…]
