### 2022-05-18 valdanchev/reproducible-data-science-python#9 @valdanchev
The differences between MyBinder and Colab are now emphasised in:
* in the [JOSE paper](https://github.com/valdanchev/reproducible-data-science-python/blob/master/JOSE_paper/paper.md)
* the learning resource website, section [How to get the most of the open learning resource?](https://valdanchev.github.io/reproducible-data-science-python/notebooks/00_textbook_outline.html#how-to-get-the-most-of-the-open-learning-resource)
* README file in the resource [repository](https://github.com/valdanchev/reproducible-data-science-python#readme)

Links to notebooks on GitHub, MyBidner, and Colab are provided in a table on the resource [website](https://valdanchev.github.io/reproducible-data-science-python/notebooks/00_textbook_outline.html#how-to-get-the-most-of-the-open-learning-resource) and in the [README file](https://github.com/valdanchev/reproducible-data-science-python#readme). Links are now removed from the paper as suggested.

The link to [Google Colaboratory](https://colab.research.google.com/notebooks/intro.ipynb#) is now functional on the [website](https://valdanchev.github.io/reproducible-data-science-python/notebooks/01_end_to_end_data_science_project.html#get-started-with-jupyter-colab-notebook).

References to the Littlest JupyterHub (also known as TLJH) are now removed, they refered to a very early in-class version.

There are now clear guidance regarding the differences between the resource website and notebooks on GitHub, MyBinder, and Colab.

Section 'Self-guided learning' in the paper discusses access and user privacy in relation to MyBinder and Colab.
