### 2022-05-18 valdanchev/reproducible-data-science-python#10 @valdanchev
The three missing hyperlinks in section [Learning resources about the A-level controversy](https://valdanchev.github.io/reproducible-data-science-python/notebooks/10_data_ethics.html#learning-resources-about-the-a-level-controversy) are now added.
