### 2022-05-18 valdanchev/reproducible-data-science-python#2 @valdanchev
Thank you for creating the issue. The error message `ModuleNotFoundError: No module named 'google'` appears when the notebook is run outside Colab, including when the notebook is executed for the build of the Jupyter Book website. The code will still execute in Colab. Added a [note](https://valdanchev.github.io/reproducible-data-science-python/notebooks/07_prediction_using_supervised_learning.html#accessing-data-from-your-google-drive) to clarify this.

In addition, added a note in the [README file](https://github.com/valdanchev/reproducible-data-science-python/blob/master/README.md) that this notebook (Prediction using Supervised Learning) and the next one (What Causes What? Introduction to Causal inference) require [access to safeguarded data](https://valdanchev.github.io/reproducible-data-science-python/notebooks/07_prediction_using_supervised_learning.html#data-understanding-society-covid-19) and private/secure storage so once data is obtained, it needs to be stored securely on Google Drive and the notebooks should be run as private Colab notebooks (as opposed to public cloud services including MyBinder).
