### 2022-05-18 #156 @valdanchev
@ShanEllis thank you for organising the review process, @TomDonoghue and @lechten thank you for your very helpful and constructive feedback and suggestions. The review process was very valuable for improving the resource, repository, and paper. Sorry for my delayed revisions.

I have now made the revisions and addressed all the feedback and comments above as well as the issues [#5](https://github.com/valdanchev/reproducible-data-science-python/issues/5), [#6](https://github.com/valdanchev/reproducible-data-science-python/issues/6),[#7](https://github.com/valdanchev/reproducible-data-science-python/issues/7),[#8](https://github.com/valdanchev/reproducible-data-science-python/issues/8),[#9](https://github.com/valdanchev/reproducible-data-science-python/issues/9), [#10](https://github.com/valdanchev/reproducible-data-science-python/issues/10). Specifically,

* In relation to code quality [#8](https://github.com/valdanchev/reproducible-data-science-python/issues/8) and code formating [#5](https://github.com/valdanchev/reproducible-data-science-python/issues/5):
  * The code now conforms to PEP 8 by running automated code formatters *Black* and *flake8* on Jupyter through [nbQA](https://github.com/nbQA-dev/nbQA), following very helpful recommendations in [#8](https://github.com/valdanchev/reproducible-data-science-python/issues/8).
 * Added a section entitled [Readability of your Python code](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#readability-of-your-python-code) in which PEP 8 style guide for Python code is introduced and key relevant recommendations for writing a readable Python code are outlined, including (i) naming conventions, (ii) maximum line length, (iii) indentation, (iv) tabs or spaces?, (v) whitespace in expressions and statements.
 * Added also a new section on code reusability named [Code reuse: creating your own functions](https://valdanchev.github.io/reproducible-data-science-python/notebooks/05_data_exploration_and_visualisation.html#code-reuse-creating-your-own-functions), related to [#8](https://github.com/valdanchev/reproducible-data-science-python/issues/8).
* In relation to content [#5](https://github.com/valdanchev/reproducible-data-science-python/issues/5), added the following sections in the notebook [Python for Data Analysis on the Cloud](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html):
  * [Running Terminal commands in Jupyter/Colab notebook](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#running-terminal-commands-in-jupyter-colab-notebook)
  * [Installing packages with pip](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#installing-packages-with-pip)
  * [Jupyter magic commands](https://valdanchev.github.io/reproducible-data-science-python/notebooks/02_python_data_science_on_the_cloud.html#jupyter-magic-commands)
* Regarding the structure of the resource GitHub repository, following the very helpful suggestions [#6](https://github.com/valdanchev/reproducible-data-science-python/issues/6), the files are now reorganised and ordered in a more structured and intuitive manner:
  * different types of content are now assigned to sub-folders `notebooks/`, `images/`, `data/`, `JOSE_paper/`;
  * notebooks are sorted in ascending order;
  * unused/unnecessary files are deleted.
* To address issues [#9](https://github.com/valdanchev/reproducible-data-science-python/issues/9) and [#6](https://github.com/valdanchev/reproducible-data-science-python/issues/6), the role of the resource website, Github, MyBinder, and Colab are now clarified and differentiated in:
  * the [JOSE paper](https://github.com/valdanchev/reproducible-data-science-python/blob/master/JOSE_paper/paper.md)
  * the learning resource website, section [How to get the most of the open learning resource?](https://valdanchev.github.io/reproducible-data-science-python/notebooks/00_textbook_outline.html#how-to-get-the-most-of-the-open-learning-resource)
  * the README in the resource [repository](https://github.com/valdanchev/reproducible-data-science-python#readme)
  * Also, links to notebooks on GitHub, MyBidner, and Colab are provided in a table on the resource [website](https://valdanchev.github.io/reproducible-data-science-python/notebooks/00_textbook_outline.html#how-to-get-the-most-of-the-open-learning-resource) and in the [README](https://github.com/valdanchev/reproducible-data-science-python#readme).
* Added a [release version identifier](https://github.com/valdanchev/reproducible-data-science-python/releases/tag/v2.0.0).
* Added notes in the [resource](https://valdanchev.github.io/reproducible-data-science-python/notebooks/07_prediction_using_supervised_learning.html#accessing-data-from-your-google-drive) and the repository [README file](https://github.com/valdanchev/reproducible-data-science-python/blob/master/README.md) addressing dependencies issues, as noted in comment [above](https://github.com/openjournals/jose-reviews/issues/156#issuecomment-1020684774) and in a [pre-existing issue](https://github.com/valdanchev/reproducible-data-science-python/issues/2)
* Regarding [Community Guidelines](https://github.com/openjournals/jose-reviews/issues/156#issuecomment-1020684774), the [README](https://github.com/valdanchev/reproducible-data-science-python/blob/master/README.md) includes now a note about how to contribute to the resource, create an issue etc.
* The first notebook [End-to-End Data Science Project](https://valdanchev.github.io/reproducible-data-science-python/notebooks/01_end_to_end_data_science_project.html) includes a [note](https://valdanchev.github.io/reproducible-data-science-python/notebooks/01_end_to_end_data_science_project.html#end-to-end-data-science-project) describing the purpose of the notebook and what is expected by the learners, as suggested by @TomDonoghue [here](https://github.com/openjournals/jose-reviews/issues/156#issuecomment-1020684774) and by @lechten [here](https://github.com/openjournals/jose-reviews/issues/156#issuecomment-1039251134).

Thank you again for the very constructive and transparent review process and feedback.
