### 2022-05-19 valdanchev/reproducible-data-science-python#7 @lechten
I just tried to generate a new PDF [for the JOSE issue](https://github.com/openjournals/jose-reviews/issues/156), which failed, probably due to the subdirectory. Do you know how to do that?
