### 2022-05-19 valdanchev/reproducible-data-science-python#7 @valdanchev
Many thanks! Just renamed the subdirectory to `paper`, which hopefully solves this; if not, another option would be to place the paper files outside the subdirectory.
