### 2022-05-19 valdanchev/reproducible-data-science-python#11 @lechten
Textbook chapters with missing output and hard-to-read warnings
This is related to your [JOSE  submission](https://github.com/openjournals/jose-reviews/issues/156).

I did not check all HTML chapters, but [that one](https://valdanchev.github.io/reproducible-data-science-python/notebooks/04_data_design_and_data_wrangling.html) shows an error when accessing data for `mobility_trends`.

Also, yellow on white is hard to read for warnings, e.g., [there](https://valdanchev.github.io/reproducible-data-science-python/notebooks/03_open_reproducible_workflows.html). Maybe you can avoid them?
