### 2022-05-19 valdanchev/reproducible-data-science-python#7 @lechten
Thanks! Now there is a [ruby exception](https://github.com/openjournals/jose-reviews/issues/156#issuecomment-1131845312)...

I have no idea whether this is a general issue or related to your paper.
