### 2022-05-19 valdanchev/reproducible-data-science-python#7 @lechten
Actually, that is probably the ":" in the title.  Maybe replace with emdash? Or quote somehow?
