### 2022-05-19 valdanchev/reproducible-data-science-python#7 @valdanchev
Totally, thanks, added quotes.
