+++ notebooks/requirements.txt
@@ -1,1 +1,1 @@
-dowhy==0.6
+pandas==1.1.3
