+++ notebooks/requirements.txt
@@ -1,1 +1,1 @@
-pandas==1.1.3
+dowhy==0.6
