### 2022-05-19 valdanchev/reproducible-data-science-python#11 @valdanchev
Many thanks! Corrected these issues. Checked also the other chapters for similar issues.
