### 2022-05-20 valdanchev/reproducible-data-science-python#7 @lechten
Thank you for the update. I asked for the PDF at JOSE...
