### 2022-05-23 valdanchev/reproducible-data-science-python#6 @TomDonoghue
Updates / reorganizations here all look good to me!
