### 2022-05-23 valdanchev/reproducible-data-science-python#5 @TomDonoghue
The updates all look good to me!
