### 2022-05-23 #156 @TomDonoghue
I have checked through the updates to the resource and everything looks good to me! I have checked of my remaining review items, and as far as I'm concerned, this is good to go!

Quick note: as things get wrapped up, I think the version number that is listed in this issue (1.0.0) needs to be updated to reflect the version number listed in the [most recent release](https://github.com/valdanchev/reproducible-data-science-python/releases/tag/v2.0.0) (I think there is a whedon command to do this).
