### 2022-05-23 #156 @valdanchev
Many thanks, @TomDonoghue. On the version number, it seems that only editors can do the update. @ShanEllis will know best how to update the version number.
