### 2022-06-21 #156 @ShanEllis
@valdanchev I'm not 100% sure, but I believe the PDF won't compile due to YAML formatting issues at the top of paper.md. Specifically, for each bullet point entry, it looks to me like a space is missing before each dash (` -`).

Sharing a link to a random accepted paper whose YAML formatting looks right to me in case helpful: https://raw.githubusercontent.com/americocunhajr/EPIDEMIC/master/paper/paper.md

Mind making these edits and trying to recompile PDF? (If this isn't the issue I'll dig deeper.)
