+++ paper/paper.md
@@ -10,14 +10,14 @@ tags:
  - machine learning
  - social networks
  - data science ethics
-date: "19 November 2021"
 authors:
  - name: Valentin Danchev
-  orcid: 0000-0002-7563-0168
-  affiliation: 1
+   orcid: 0000-0002-7563-0168
+   affiliation: 1
 affiliations:
  - name: Department of Sociology, University of Essex, United Kingdom
-  index: 1
+   index: 1
+date: 21 June 2022   
 bibliography: paper.bib
 ---
 
