+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
     orcid: 0000-0002-7563-0168
     affiliation: 1
 affiliations:
-  - name: Department of Sociology, University of Essex, United Kingdom
+  - name: 'Department of Sociology, University of Essex, United Kingdom'
     index: 1
 date: 21 June 2022   
 bibliography: paper.bib
