+++ paper/paper.bib
@@ -2,7 +2,6 @@
   journal = {Harvard Data Science Review},
   doi = {10.1162/99608f92.cb0fa8d2},
   note = {https://hdsr.mitpress.mit.edu/pub/e69066t4},
-  publisher = {},
   title = {Interleaving Computational and Inferential Thinking: Data Science for Undergraduates at Berkeley},
   url = {https://hdsr.mitpress.mit.edu/pub/e69066t4},
   author = {Adhikari, Ani and DeNero, John and Jordan, Michael I.},
@@ -12,7 +11,7 @@
   day = {30},
 }
 
-@BOOK{NAP2018,
+@book{NAP2018,
   author    = {National Academies of Sciences, Engineering, and Medicine},
   title     = {Data Science for Undergraduates: Opportunities and Options},
   isbn      = {978-0-309-47559-4},
@@ -82,7 +81,7 @@
   publisher={Nature Publishing Group},
 }
 
-@Book{Lau2021,
+@book{Lau2021,
   title     = {Principles and Techniques of Data Science},
   year      = {2021},
   author    = {Sam Lau and Joey Gonzalez and Deb Nolan},
@@ -109,7 +108,7 @@
   pages = {11-15},
   place = {Pasadena, CA, United States},
   year = {2008},
-  editor =    {Gael Varoquaux and Travis Vaught and Jarrod Millman},
+  editor = {Gael Varoquaux and Travis Vaught and Jarrod Millman},
 }
 
 @article{Pedregosa2011scikit, 
@@ -142,7 +141,7 @@
   pages = {3021},
   author = {Michael L. Waskom},
   title = {seaborn: statistical data visualization},
-  journal = {Journal of Open Source Software}
+  journal = {Journal of Open Source Software},
  }
 
 @inproceedings{Seabold2010statsmodels,
@@ -173,7 +172,7 @@
   year={2020},
   eprint={2004.04145},
   archivePrefix={arXiv},
-  primaryClass={cs.CR}
+  primaryClass={cs.CR},
 }
 
 @article{knuth1984literate,
@@ -185,7 +184,7 @@
   number={2},
   pages={97--111},
   year={1984},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
 }
 
 @article{Rule2019,
@@ -212,14 +211,14 @@
   month={Dec},
 }
 
-@InProceedings{ project_jupyter-proc-scipy-2018,
-  author    = { {P}roject {J}upyter and {M}atthias {B}ussonnier and {J}essica {F}orde and {J}eremy {F}reeman and {B}rian {G}ranger and {T}im {H}ead and {C}hris {H}oldgraf and {K}yle {K}elley and {G}ladys {N}alvarte and {A}ndrew {O}sheroff and {M} {P}acer and {Y}uvi {P}anda and {F}ernando {P}erez and {B}enjamin {R}agan-{K}elley and {C}arol {W}illing },
-  title     = { {B}inder 2.0 - {R}eproducible, interactive, sharable environments for science at scale },
-  booktitle = { {P}roceedings of the 17th {P}ython in {S}cience {C}onference },
-  pages     = { 113 - 120 },
-  year      = { 2018 },
-  editor    = { {F}atih {A}kici and {D}avid {L}ippa and {D}illon {N}iederhut and {M} {P}acer },
-  doi       = { 10.25080/Majora-4af1f417-011 }
+@InProceedings{project_jupyter-proc-scipy-2018,
[…]
