+++ paper/paper.bib
@@ -211,7 +211,7 @@
   month={Dec},
 }
 
-@InProceedings{project_jupyter-proc-scipy-2018,
+@InProceedings{project_jupyter_proc_scipy_2018,
   author    = {{P}roject {J}upyter and {M}atthias {B}ussonnier and {J}essica {F}orde and {J}eremy {F}reeman and {B}rian {G}ranger and {T}im {H}ead and {C}hris {H}oldgraf and {K}yle {K}elley and {G}ladys {N}alvarte and {A}ndrew {O}sheroff and {M} {P}acer and {Y}uvi {P}anda and {F}ernando {P}erez and {B}enjamin {R}agan-{K}elley and {C}arol {W}illing},
   title     = {{B}inder 2.0 - {R}eproducible, interactive, sharable environments for science at scale},
   booktitle = {{P}roceedings of the 17th {P}ython in {S}cience {C}onference},
