### 2022-06-23 #156 @arfon
This is probably because you have three `paper.md` files in your repo, and @whedon will just pick the first by default:

```
1. tmp/156/_build/html/_sources/JOSE_paper/paper.md
2. tmp/156/_build/html/_sources/paper/paper.md
3. tmp/156/paper/paper.md
```

The third one (at `tmp/156/paper/paper.md`) seems to compile fine. I would suggest either removing or renaming the first two (to not be called `paper.md`).
