### 2022-06-23 valdanchev/reproducible-data-science-python#7 @lechten
Thank you for your patience. With the new PDF, everything seems fine to me. Closing this.
