+++ paper/paper.bib
@@ -2,7 +2,7 @@
   journal = {Harvard Data Science Review},
   doi = {10.1162/99608f92.cb0fa8d2},
   note = {https://hdsr.mitpress.mit.edu/pub/e69066t4},
-  title = {Interleaving Computational and Inferential Thinking: Data Science for Undergraduates at Berkeley},
+  title = {{I}nterleaving {C}omputational and {I}nferential {T}hinking: {D}ata {S}cience for {U}ndergraduates at {B}erkeley},
   url = {https://hdsr.mitpress.mit.edu/pub/e69066t4},
   author = {Adhikari, Ani and DeNero, John and Jordan, Michael I.},
   date = {2021-04-30},
@@ -13,7 +13,7 @@
 
 @book{NAP2018,
   author    = {National Academies of Sciences, Engineering, and Medicine},
-  title     = {Data Science for Undergraduates: Opportunities and Options},
+  title     = {{D}ata {S}cience for {U}ndergraduates: {O}pportunities and {O}ptions},
   isbn      = {978-0-309-47559-4},
   doi       = {10.17226/25104},
   url       = {https://www.nap.edu/catalog/25104/data-science-for-undergraduates-opportunities-and-options},
@@ -25,19 +25,18 @@
 @inproceedings{Danyluk2021, 
   author = {Danyluk, Andrea and Leidig, Paul and McGettrick, Andrew and Cassel, Lillian and Doyle, Maureen and Servin, Christian and Schmitt, Karl and Stefik, Andreas}, 
   doi = {10.1145/3408877.3432586},
-  title = {Computing Competencies for Undergraduate Data Science Programs: An ACM Task Force Final Report}, year = {2021}, isbn = {9781450380621}, 
+  title = {{C}omputing {C}ompetencies for {U}ndergraduate {D}ata {S}cience {P}rograms: An {ACM} {T}ask {F}orce {F}inal {R}eport}, year = {2021}, isbn = {9781450380621}, 
   publisher = {Association for Computing Machinery}, 
   address = {New York, NY, USA}, url = {https://doi.org/10.1145/3408877.3432586}, doi = {10.1145/3408877.3432586}, 
   booktitle = {Proceedings of the 52nd ACM Technical Symposium on Computer Science Education}, 
   pages = {1119–1120},
   numpages = {2},
-  keywords = {data science, curriculum},
   location = {Virtual Event, USA}, 
   series = {SIGCSE '21},
 }
 
 @article{Arnold2019turing,
-  title={The turing Way: a handbook for reproducible data science},
+  title={The {T}uring {W}ay: a {H}andbook for {R}eproducible {D}ata {S}cience},
   author={Arnold, Becky and Bowler, Louise and Gibson, Sarah and Herterich, Patricia and Higman, Rosie and Krystalli, Anna and Morley, Alexander and O'Reilly, Martin and Whitaker, Kirstie and others},
   doi = {10.5281/zenodo.3233853}, 
   journal={Zenodo},
@@ -49,7 +48,7 @@
   author = {Stoudt, Sara AND Vásquez, Váleri N. AND Martinez, Ciera C.},
   journal = {PLOS Computational Biology},
   publisher = {Public Library of Science},
-  title = {Principles for data analysis workflows},
+  title = {Principles for {D}ata {A}nalysis {W}orkflows},
   year = {2021},
   month = {03},
   volume = {17},
@@ -59,7 +58,7 @@
 }
 
 @conference{Kluyver2016jupyter, 
-  Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows}, 
[…]
