+++ paper/paper.bib
@@ -124,7 +124,7 @@
   author    = {Wes McKinney},
   title     = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
   doi = {10.25080/Majora-92bf1922-00a},
-  booktitle = {Proceedings of the 9th Python in Science Conference},
+  booktitle = {Proceedings of the 9th {P}ython in {S}cience {C}onference},
   pages     = {51-56},
   year      = {2010},
   editor    = {Stefan van der Walt and Jarrod Millman},
@@ -147,7 +147,7 @@
   title={{S}tatsmodels: {E}conometric and {S}tatistical {M}odeling with {P}ython},
   author={Seabold, Skipper and Perktold, Josef},
   doi = {10.25080/Majora-92bf1922-011},  
-  booktitle={9th Python in Science Conference},
+  booktitle={{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
   year={2010},
 }
 
