+++ paper/paper.bib
@@ -212,7 +212,7 @@
 
 @InProceedings{project_jupyter_proc_scipy_2018,
   author    = {{P}roject {J}upyter and {M}atthias {B}ussonnier and {J}essica {F}orde and {J}eremy {F}reeman and {B}rian {G}ranger and {T}im {H}ead and {C}hris {H}oldgraf and {K}yle {K}elley and {G}ladys {N}alvarte and {A}ndrew {O}sheroff and {M} {P}acer and {Y}uvi {P}anda and {F}ernando {P}erez and {B}enjamin {R}agan-{K}elley and {C}arol {W}illing},
-  title     = {{B}inder 2.0 - {R}eproducible, {I}nteractive, {S}harable {E}nvironments for {S}cience at {S}cale},
+  title     = {{B}inder 2.0 — {R}eproducible, {I}nteractive, {S}harable {E}nvironments for {S}cience at {S}cale},
   booktitle = {{P}roceedings of the 17th {P}ython in {S}cience {C}onference},
   pages     = {113 - 120},
   year      = {2018},
@@ -228,5 +228,5 @@
   series={Chapman \& Hall/CRC Data Science Series},
   url={https://datasciencebook.ca/},
   year={2022},
-  publisher={Taylor \& Francis Limited},
+  publisher={Chapman and Hall/CRC},
 }
