+++ paper/paper.bib
@@ -228,5 +228,5 @@
   series={Chapman \& Hall/CRC Data Science Series},
   url={https://datasciencebook.ca/},
   year={2022},
-  publisher={Chapman \and Hall/CRC},
+  publisher={Chapman \& Hall/CRC},
 }
