+++ paper/paper.bib
@@ -58,7 +58,7 @@
 }
 
 @conference{Kluyver2016jupyter, 
-  Title = {{J}upyter {N}otebooks – {A} {P}ublishing {F}ormat for {R}eproducible {C}omputational {W}orkflows}, 
+  Title = {{J}upyter {N}otebooks — {A} {P}ublishing {F}ormat for {R}eproducible {C}omputational {W}orkflows}, 
   Author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing}, 
   doi = {10.3233/978-1-61499-649-1-87},
   Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas}, 
@@ -202,7 +202,7 @@
 }
 
 @misc{guzharina2020, 
-  title={We {D}ownloaded 10,000,000 {J}upyter {N}otebooks from {G}it{H}ub – {T}his is {W}hat we {L}earned}, 
+  title={We {D}ownloaded 10,000,000 {J}upyter {N}otebooks from {G}it{H}ub — {T}his is {W}hat we {L}earned}, 
   url={https://blog.jetbrains.com/datalore/2020/12/17/we-downloaded-10-000-000-jupyter-notebooks-from-github&-this-is-what-we-learned/}, 
   journal={The JetBrains Datalore Blog}, 
   author={Alena Guzharina}, 
