+++ paper/paper.bib
@@ -188,7 +188,7 @@
 
 @article{Rule2019,
   doi = {10.1371/journal.pcbi.1007007},
-  author = {Rule, Adam AND Birmingham, Amanda AND Zuniga, Cristal AND Altintas, Ilkay AND Huang, Shih-Cheng AND Knight, Rob AND Moshiri, Niema AND Nguyen, Mai H. AND Rosenthal, Sara Brin AND Pérez, Fernando AND Rose, Peter W.},
+  author = {Rule, Adam and Birmingham, Amanda and Zuniga, Cristal and Altintas, Ilkay and Huang, Shih-Cheng and Knight, Rob and Moshiri, Niema and Nguyen, Mai H. and Rosenthal, Sara Brin and Pérez, Fernando and Rose, Peter W.},
   journal = {PLOS Computational Biology},
   publisher = {Public Library of Science},
   title = {Ten {S}imple {R}ules for {W}riting and {S}haring {C}omputational {A}nalyses in {J}upyter {N}otebooks},
