+++ paper/paper.bib
@@ -64,7 +64,7 @@
   Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas}, 
   Editor = {F. Loizides and B. Schmidt}, 
   Organization = {IOS Press}, 
-  Pages = {87 - 90}, 
+  Pages = {87--90}, 
   Year = {2016},
 }
 
@@ -93,7 +93,7 @@
   journal = {CHANCE},
   volume = {32},
   number = {1},
-  pages = {42-49},
+  pages = {42--49},
   year  = {2019},
   publisher = {Taylor & Francis},
   doi = {10.1080/09332480.2019.1579578},
