### 2022-07-21 #156 @labarba
@valdanchev — As this review is finalized, please make an archive of the target repository on Zenodo, and report the archive DOI here. We request that authors edit the metadata of the Zenodo deposit so title and author list match the JOSE paper.
