+++ _sources/JOSE_paper/paper.md
@@ -1,137 +1,0 @@
----
-title: An open learning resource on Reproducible Data Science with Python
-tags:
-- Python
-- Jupyter notebook
-- reproducible workflow
-- open science
-- real-world social data
-- exploratory data analysis
-- machine learning
-- social networks
-- data science ethics
-date: "19 November 2021"
-authors:
-- name: Valentin Danchev
-  orcid: 0000-0002-7563-0168
-  affiliation: 1
-affiliations:
-- name: Department of Sociology, University of Essex, United Kingdom
-  index: 1
-bibliography: paper.bib
----
-
-# Summary
-
-This paper describes a computational learning resource on Reproducible Data Science with Python. The resource provides an accessible, hands-on introduction to data science 
-techniques, skills, and workflows necessary to perform open, reproducible, and ethical data analysis. By using research problems of real-world relevance (such as vaccine hesitancy and the impact of COVID-19 
-lockdown measures on human mobility) and real-world social data (including anonymised mobility data from digital sources and recent COVID-19 survey data), the resource encourages students to use 
-open-source tools and coding to learn from diverse and large social data sources.
-
-The learning resource aims to minimise barriers to entry for students from social sciences, public health, and related fields. With no software installation and setup requirements, students can start coding 
-from their web browser using free and open-source software (FOSS), including the Python programming language, Jupyter notebook, and Markdown. Through real-world data applications, students are introduced 
-to the open source Python ecosystem of libraries for data science—including `pandas` [@Mckinney2010], `seaborn` [@Waskom2021], `scikit-learn` [@Pedregosa2011scikit], `statsmodels` [@Seabold2010statsmodels], 
-and `networkX` [@Hagberg2008networkx]—and learn about open and reproducible workflow, data wrangling, data exploration and visualization, pattern discovery (e.g., clustering), prediction and machine learning, 
-causal inference, network analysis, and data ethics.
-
-# Statement of Need
-
-The importance of data science education for drawing conclusions from diverse data sources in a transparent, reproducible, and ethical manner and for building data literacy skills 
-is now widely recognised [@NAP2018; @Danyluk2021; @Adhikari2021; @Arnold2019turing; @Stoudt2021]. For example, the National Academies of Science, Engineering, and Medicine’s 2018 'Data Science for Undergraduates' 
-report [@NAP2018] identified the importance of workflow, reproducibility, and ethical problem solving in data science undergraduate curriculum. A key challenge ahead for the community of data science educators, and particularly those at the 
-intersection of data science and social sciences, is how to democratise the knowledge and skills needed for conducting reproducible data science and to engage students—with little to no programming 
-experience and from diverse social and academic backgrounds—in accessible, inclusive, and cross-disciplinary data science learning. To address this and related challenges, successful examples of data 
-science curriculum [e.g., @Adhikari2021, @timbers2022data] have been developed. The learning resource presented in this paper addresses the data-science democratisation challenge through the use of down-to-earth research 
-questions and real-world social data sets about the COVID-19 pandemic. Both real-world questions and data can encourage students to engage with hands-on computation, data science techniques, reproducible 
-data analysis workflow, and data ethics. The democratisation of reproducible and ethical data science is important for empowering students (and citizens) to use, analyse, and learn from large data sets as 
-well as to critically evaluate data, models, and their social impact.
-
-# Learning objectives
-
-The learning resource aims to enable students to:
-
-* Freely use open-source computational tools—e.g., Python, Jupyter Notebook, Markdown—for data analysis.
-* Build a transparent and reproducible research workflow.
-* Perform data analysis and critically interpret and openly communicate research process, objects, and results.
[…]
