+++ _sources/paper/paper.md
@@ -1,23 +1,24 @@
 ---
-title: 'Reproducible Data Science with Python: An open learning resource'
+title: 'Reproducible Data Science with Python: An Open Learning Resource'
 tags:
-- Python
-- Jupyter notebook
-- reproducible workflow
-- open science
-- real-world social data
-- exploratory data analysis
-- machine learning
-- social networks
-- data science ethics
-date: "19 November 2021"
+  - data science
+  - Python
+  - Jupyter notebook
+  - reproducible workflow
+  - open science
+  - real-world social data
+  - exploratory data analysis
+  - machine learning
+  - social networks
+  - data science ethics
 authors:
-- name: Valentin Danchev
-  orcid: 0000-0002-7563-0168
-  affiliation: 1
+  - name: Valentin Danchev
+    orcid: 0000-0002-7563-0168
+    affiliation: 1
 affiliations:
-- name: Department of Sociology, University of Essex, United Kingdom
-  index: 1
+  - name: Department of Sociology, University of Essex, United Kingdom
+    index: 1
+date: 21 June 2022
 bibliography: paper.bib
 ---
 
@@ -34,14 +35,14 @@ to the open source Python ecosystem of libraries for data science—including `p
 and `networkX` [@Hagberg2008networkx]—and learn about open and reproducible workflow, data wrangling, data exploration and visualization, pattern discovery (e.g., clustering), prediction and machine learning, 
 causal inference, network analysis, and data ethics.
 
-# Statement of Need
+# Statement of need
 
 The importance of data science education for drawing conclusions from diverse data sources in a transparent, reproducible, and ethical manner and for building data literacy skills 
 is now widely recognised [@NAP2018; @Danyluk2021; @Adhikari2021; @Arnold2019turing; @Stoudt2021]. For example, the National Academies of Science, Engineering, and Medicine’s 2018 'Data Science for Undergraduates' 
 report [@NAP2018] identified the importance of workflow, reproducibility, and ethical problem solving in data science undergraduate curriculum. A key challenge ahead for the community of data science educators, and particularly those at the 
 intersection of data science and social sciences, is how to democratise the knowledge and skills needed for conducting reproducible data science and to engage students—with little to no programming 
 experience and from diverse social and academic backgrounds—in accessible, inclusive, and cross-disciplinary data science learning. To address this and related challenges, successful examples of data 
-science curriculum [e.g., @Adhikari2021, @timbers2022data] have been developed. The learning resource presented in this paper addresses the data-science democratisation challenge through the use of down-to-earth research 
+science curriculum [e.g., @Adhikari2021; @timbers2022data] have been developed. The learning resource presented in this paper addresses the data-science democratisation challenge through the use of down-to-earth research 
 questions and real-world social data sets about the COVID-19 pandemic. Both real-world questions and data can encourage students to engage with hands-on computation, data science techniques, reproducible 
 data analysis workflow, and data ethics. The democratisation of reproducible and ethical data science is important for empowering students (and citizens) to use, analyse, and learn from large data sets as 
 well as to critically evaluate data, models, and their social impact.
@@ -104,7 +105,7 @@ notebook does not guarantee reproducibility [@guzharina2020], it fosters reprodu
 # Experience of use
 
 The learning resource can be used for different modes of learning, including a semester-long module, training workshop, and independent study. So far, elements of the learning resource have been 
-used in a 10-week module delivered in the Spring term of the academic year 2020–2021 to third year undergraduate students at the Department of Sociology, University of Essex. More recently, 
+used in a 10-week module delivered in the Spring term of the academic years 2020–21 and 2021–22 to third year undergraduate students at the Department of Sociology, University of Essex. More recently, 
[…]
