### 2022-07-24 #156 @valdanchev
@labarba and @ShanEllis — I've archived the resource repository on Zenodo. The archived version is v2.1.1.

Here are:
URL: https://zenodo.org/record/6895578#.Yt1gSy8w3Ls
DOI: 10.5281/zenodo.6895578
DOI badge [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6895578.svg)](https://doi.org/10.5281/zenodo.6895578)

I edited the metadata of the Zenodo repository to match the JOSE paper (I revised/shortened the paper title during the revision and the Zenodo metadata matches the revised title).
