### 2022-08-22 #156 @valdanchev
@labarba & @ShanEllis — I think all should work now. There was a problem similar to the one Arfon identified before — a redundant paper.md file was generated after an update just before archiving the repository on Zenodo, — which is now corrected. The pdf can be generated again and hopefully, the other commands would work too.
