### 2022-10-23 #156 @labarba
hi @valdanchev  — it looks like this one slipped through the cracks, sorry (I have total email overload). On a quick browse, I noticed that the repo has as latest tag [v2.1.1](https://github.com/valdanchev/reproducible-data-science-python/releases/tag/v2.1.1), while the version here is noted as v2.0.0. Meanwhile, you have a lot of newer commits. (Unfortunately, you do not have meaningful commit messages, as all are titled "Updates"—I do urge you to revisit this practice!!)

Can you make a tagged release with all the latest changes, corresponding to the reviewed and revised JOSE submission? Then report the version number here and we will update it.
