### 2022-10-24 42be31c
fixed broken URLs in Intro, updated affiliations & acknowledgements
