+++ paper/paper.md
@@ -14,10 +14,14 @@ tags:
 authors:
   - name: Valentin Danchev
     orcid: 0000-0002-7563-0168
-    affiliation: 1
+    affiliation: "1, 2"
 affiliations:
   - name: Department of Sociology, University of Essex, United Kingdom
     index: 1
+affiliations:
+  - name: School of Business and Management, Queen Mary University of London, United Kingdom
+    index: 2
+
 date: 21 June 2022
 bibliography: paper.bib
 ---
