### 2022-10-24 #156 @valdanchev
@labarba - many thanks, this is very helpful! I have now made a new tagged release corresponding to the JOSE publication (with added informative commit messages) and updated the Zenodo archive, details are below. Let me know if any additional changes are needed. Thank you again.

Latest repo version: [v2.1.2](https://github.com/valdanchev/reproducible-data-science-python/releases/tag/v2.1.2)
Zenodo DOI: [10.5281/zenodo.7244097](https://zenodo.org/record/7244097)
