### 2022-10-24 #156 @labarba
hi @valdanchev — One last thing: We request that authors edit the metadata of the Zenodo deposit so title and author list match the JOSE paper. It's just cleaner that way as readers see these as part of the "same scholarly object." Could you do that?
