### 2022-10-24 #156 @valdanchev
hi @labarba — thank you, more than happy to do that, just updated the metadata of the deposit.
