### 2022-10-24 #156 @labarba
Congratulations, @valdanchev, your JOSE paper is accepted! 🚀

Huge thanks to our editor: @ShanEllis and the reviewers: @TomDonoghue, @lechten — your contribution makes this adventure possible 🙏
