### 2022-10-24 #156 @valdanchev
Great!! Thank you so much @labarba, @ShanEllis, @TomDonoghue, and @lechten for all of your work — the review process has really improved the learning resource!
